Fix issue when swapping bytes of signed integers in SPARC architecture

+++ src/client-agent/receiver-win.c
@@ -95,8 +95,8 @@ void *receiver_thread(__attribute__((unused)) void *none)
 
                 recv_b = recv(agt->sock, buffer, length, MSG_WAITALL);
 
-                if (recv_b != length) {
-                    merror("Incorrect message size from manager: expecting %u, got %zd", length, recv_b);
+                if (recv_b != (ssize_t)length) {
+                    merror("Incorrect message size from manager: expecting %u, got %d", length, (int)recv_b);
                     break;
                 }
             } else {
